correctly run on gpu

+++ CRL/ENVELOPE/meta.py
@@ -17,13 +17,13 @@ Tensor = FloatTensor
 
 class MetaAgent(object):
 	'''
-	(1) act: how to sample an action to examine the learning 
+	(1) act: how to sample an action to examine the learning
 		outcomes or explore the environment;
 	(2) memorize: how to store observed observations in order to
-		help learing or establishing the empirical model of the 
+		help learing or establishing the empirical model of the
 		enviroment;
-	(3) learn: how the agent learns from the observations via 
-		explicitor implicit inference, how to optimize the policy 
+	(3) learn: how the agent learns from the observations via
+		explicitor implicit inference, how to optimize the policy
 		model.
 	'''
 	def __init__(self, model, args, is_train=False):
@@ -52,8 +52,8 @@ class MetaAgent(object):
 
 		if self.is_train: self.model.train()
 		if use_cuda: self.model.cuda()
-		
-		
+
+
 	def act(self, state, preference=None):
 		# random pick a preference if it is not specified
 		if preference is None:
@@ -69,7 +69,7 @@ class MetaAgent(object):
 		state = torch.from_numpy(state).type(FloatTensor)
 
 		_, Q = self.model(
-				Variable(state.unsqueeze(0)), 
+				Variable(state.unsqueeze(0)),
 				Variable(preference.unsqueeze(0)))
 
 		Q = Q.view(-1, self.model.reward_size)
@@ -82,7 +82,7 @@ class MetaAgent(object):
 		if self.is_train and (len(self.trans_mem) < self.batch_size or \
 					torch.rand(1)[0] < self.epsilon):
 				action = np.random.choice(self.model.action_size, 1)[0]
-			
+
 		return action
 
 
@@ -93,19 +93,19 @@ class MetaAgent(object):
 							torch.from_numpy(next_state).type(FloatTensor), 	# next state
 							torch.from_numpy(reward).type(FloatTensor), 		# reward
 							terminal))								# terminal
-		
+
 		# randomly produce a preference for calculating priority
 		# preference = self.keep_preference
 		preference = torch.randn(self.model.reward_size)
 		preference = (torch.abs(preference) / torch.norm(preference, p=1)).type(FloatTensor)
 		state = torch.from_numpy(state).type(FloatTensor)
-		
+
 		_, q  = self.model(Variable(state.unsqueeze(0), volatile=True),
 						  Variable(preference.unsqueeze(0), volatile=True))
-	
+
 		q = q[0, action].data
 		wq = preference.dot(q)
-		
+
 		wr = preference.dot(torch.from_numpy(reward).type(FloatTensor))
 		if not terminal:
 			next_state = torch.from_numpy(next_state).type(FloatTensor)
@@ -137,7 +137,7 @@ class MetaAgent(object):
 				p=pri/pri.sum()
 			)
 		return [pop[i] for i in inds]
-	
+
 
 	def actmsk(self, num_dim, index):
 		mask = ByteTensor(num_dim).zero_()
@@ -167,7 +167,7 @@ class MetaAgent(object):
 			# # preference_batch = np.array([[0.8,0.2], [0.2, 0.8]])
 			preference_batch = np.abs(preference_batch) / \
 								np.linalg.norm(preference_batch, ord=1, axis=1, keepdims=True)
-			# preference_batch = np.random.dirichlet(np.ones(self.model.reward_size), size=self.weight_num)								
+			# preference_batch = np.random.dirichlet(np.ones(self.model.reward_size), size=self.weight_num)
 			preference_batch = torch.from_numpy(preference_batch.repeat(self.batch_size, axis=0)).type(FloatTensor)
 
 			__, Q    = self.model(Variable(torch.cat(state_batch, dim=0)),
@@ -176,23 +176,23 @@ class MetaAgent(object):
 			HQ, _    = self.model(Variable(torch.cat(next_state_batch, dim=0)),
 								  Variable(preference_batch), w_num=self.weight_num)
 
-			
+
 			nontmlmask = self.nontmlinds(terminal_batch)
-			Estimate_Q = Variable(torch.zeros(self.batch_size*self.weight_num, 
-											  self.model.reward_size))
+			Estimate_Q = Variable(torch.zeros(self.batch_size*self.weight_num,
+											  self.model.reward_size).type(FloatTensor))
 			Estimate_Q[nontmlmask] = self.gamma * HQ[nontmlmask]
 			Estimate_Q += Variable(torch.cat(reward_batch, dim=0))
 
 			self.optimizer.zero_grad()
 			action_mask = Variable(torch.cat(action_batch, dim=0))
 			action_mask = action_mask.unsqueeze(2).repeat(1,1,self.model.reward_size)
-			loss = torch.sum((Q.masked_select(action_mask) - Estimate_Q.view(-1)).pow(2))			
+			loss = torch.sum((Q.masked_select(action_mask) - Estimate_Q.view(-1)).pow(2))
 			report_loss = loss.data[0]/(self.batch_size*self.weight_num)
 			loss.backward()
 			self.optimizer.step()
 
 			return	report_loss
-		
+
 		return 1.0
 
 	def reset(self):
@@ -203,5 +203,3 @@ class MetaAgent(object):
 
 	def save(self, save_path, model_name):
 		torch.save(self.model, "{}{}.pkl".format(save_path, model_name))
-
-
